+++ src/test/java/com/ontotext/refine/client/SecuredExportIntegrationTest.java
@@ -1,6 +1,7 @@
 package com.ontotext.refine.client;
 
 import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
@@ -8,6 +9,8 @@ import com.ontotext.refine.client.command.RefineCommands;
 import com.ontotext.refine.client.command.create.CreateProjectResponse;
 import com.ontotext.refine.client.command.delete.DeleteProjectResponse;
 import com.ontotext.refine.client.command.rdf.ExportRdfResponse;
+import com.ontotext.refine.client.command.rdf.GraphDbSparqlBasedRdfExportCommand;
+import com.ontotext.refine.client.command.rdf.OutputType;
 import com.ontotext.refine.client.command.rdf.ResultFormat;
 import com.ontotext.refine.client.util.RdfTestUtils;
 import java.io.ByteArrayInputStream;
@@ -76,13 +79,36 @@ class SecuredExportIntegrationTest extends CommandIntegrationTest {
     String query = IOUtils.resourceToString(
         "/integration/netherlands_restaurants_construct_query.sparql",
         StandardCharsets.UTF_8);
-    return RefineCommands
-        .exportRdfUsingSparql()
-        .setProject(projectId)
-        .setFormat(ResultFormat.TURTLE)
-        .setQuery(query)
-        .setRepository(REPO_NAME)
-        .build()
-        .execute(getClient());
+    TestGraphDbSparqlBasedRdfExportCommand command = new TestGraphDbSparqlBasedRdfExportCommand(
+        projectId, "#project_placeholder#", query, ResultFormat.TURTLE, REPO_NAME, null);
+    return command.execute(getClient());
+  }
+
+  // hack for the proxy in order to keep the test and the logic around it, because it is relevant
+  private class TestGraphDbSparqlBasedRdfExportCommand extends GraphDbSparqlBasedRdfExportCommand {
+
+    private TestGraphDbSparqlBasedRdfExportCommand(
+        String project,
+        String placeholder,
+        String query,
+        ResultFormat format,
+        String repository,
+        OutputType output) {
+      super(project, placeholder, query, format, repository, output);
+
+      // verifies the input parameters, also Sonar test coverage
+      GraphDbSparqlBasedRdfExportCommand.Builder builder = RefineCommands
+          .exportRdfUsingSparql()
+          .setProject(project)
+          .setFormat(ResultFormat.TURTLE)
+          .setQuery(query)
+          .setRepository(REPO_NAME);
+      assertDoesNotThrow(() -> builder.build());
+    }
+
+    @Override
+    public String endpoint() {
+      return "/repositories/{repo}";
+    }
   }
 }